webapp/backend/requirements.txt
fastapi>=0.109.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.22
aiofiles>=23.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
reportlab>=4.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
